LPS-60972 The classesDir sourceSet configuration is specific to Liferay
